+++ arm_attributes.h
@@ -18,25 +18,29 @@
 #if defined (__CC_ARM)
 	/*** ARM REAL VIEW ***/
 	// Declared as qualifier
-	#define __WEAK				__weak								//!< \b Weak attribute
-	#define __IRQ				__irq								//!< \b Interrupt attribute
+	#define __WEAK				__weak									//!< \b Weak attribute
+	#define __IRQ				__irq									//!< \b Interrupt attribute
 	// Declared as attribute
-	#define ATTR__(...)			__attribute__((__VA_ARGS__))		//!< Macro to define one or multiple attribute(s) \b ... for a declaration
+	#define ATTR__(...)			__attribute__((__VA_ARGS__))			//!< Macro to define one or multiple attribute(s) \b ... for a declaration
 
-	#define ALIGN__(n)			__attribute__((align(n)))			//!< \b Align attribute padded to \b n
+	#define ALIGN__(n)			__attribute__((align(n)))				//!< \b Align attribute padded to \b n
 	#define COLD__
-	#define DEPRECATED__		__attribute__((deprecated))			//!< \b Deprecated attribute
+	#define DEPRECATED__		__attribute__((deprecated))				//!< \b Deprecated attribute
 	#define HOT__
-	#define INLINE__			__attribute__((always_inline))		//!< \b Always \b inline attribute
+	#define INLINE__			__attribute__((always_inline))			//!< \b Always \b inline attribute
 	#define NONNULL__
 	#define NONNULLX__(...)
-	#define NORETURN__			__attribute__((noreturn))			//!< \b No \b return attribute
-	#define PACK__				__attribute__((packed))				//!< \b Packed attribute
-	#define PURE__				__attribute__((pure))				//!< \b Pure attribute
-	#define SECTION__(s)		__attribute__((section(#s)))		//!< \b Section attribute to place declaration into section \b s
-	#define USED__				__attribute__((used))				//!< \b Used attribute ensures declaration won't be removed by garbage collector
+	#define NORETURN__			__attribute__((noreturn))				//!< \b No \b return attribute
+	#define PACK__				__attribute__((packed))					//!< \b Packed attribute
+	#define PURE__				__attribute__((pure))					//!< \b Pure attribute
+	#define SECTION__(s)		__attribute__((section(#s)))			//!< \b Section attribute to place declaration into section \b s
+	#define USED__				__attribute__((used))					//!< \b Used attribute ensures declaration won't be removed by garbage collector
 	// Common mixed attributes
 	#define NONNULL_INLINE__	ATTR__(nonnull, always_inline)			//!< \b Ensures declaration won't be removed by garbage collector
+
+	#if (!defined(__OPTIMIZE__))
+	#define __NOOPT__													//!< \b No \b Optimizations attribute
+	#endif
 
 #elif defined(__ARMCC_VERSION) && (__ARMCC_VERSION >= 6010050)
 	/*** ARM COMPILER TOOLCHAIN ***/
@@ -61,6 +65,10 @@
 	// Common mixed attributes
 	#define NONNULL_INLINE__	ATTR__(nonnull, always_inline)			//!< \b Ensures declaration won't be removed by garbage collector
 
+	#if (!defined(__OPTIMIZE__))
+	#define __NOOPT__													//!< \b No \b Optimizations attribute
+	#endif
+
 #elif defined (__GNUC__)
 	/*** GCC ***/
 	// Declared as qualifier
@@ -83,6 +91,10 @@
 	#define USED__				__attribute__((used))					//!< \b Used attribute ensures declaration won't be removed by garbage collector
 	// Common mixed attributes
 	#define NONNULL_INLINE__	ATTR__(nonnull, always_inline)			//!< \b Ensures declaration won't be removed by garbage collector
+
+	#if (!defined(__OPTIMIZE__))
+	#define __NOOPT__													//!< \b No \b Optimizations attribute
+	#endif
 
 #elif defined (__ICCARM__)
 	/*** IAR ***/
@@ -107,6 +119,8 @@
 	// Common mixed attributes
 	#define NONNULL_INLINE__
 
+	//#define __NOOPT__													//!< \b No \b Optimizations attribute not identified on IAR toolchain
+
 #elif defined (__TMS470__)
 	/*** TEXAS INSTRUMENTS ***/
 	// Declared as qualifier
@@ -129,6 +143,8 @@
 	#define USED__				__attribute__((used))					//!< \b Used attribute ensures declaration won't be removed by garbage collector
 	// Common mixed attributes
 	#define NONNULL_INLINE__
+
+	//#define __NOOPT__													//!< \b No \b Optimizations attribute not identified on TI toolchain
 
 #elif defined (__TASKING__)
 	/*** TASKING ***/
@@ -153,12 +169,27 @@
 	// Common mixed attributes
 	#define NONNULL_INLINE__
 
+	//#define __NOOPT__													//!< \b No \b Optimizations attribute not identified on TASKING toolchain
+
 #elif defined (__CSMC__)
 	/*** COSMIC ***/
-	#error No attribute defined yet for Cosmic compiler.
+	#error No attribute defined yet with Cosmic compiler.
 
 #else
 	#error Unknown compiler. Attributes will not be recognized.
+#endif
+
+
+/*** INLINES WORKAROUND WHEN OPTIMIZATION LEVEL SET TO NONE (GCC like toolchains) ***/
+#if defined(__NOOPT__)
+#ifdef __INLINE
+#undef __INLINE
+#endif
+#define __INLINE			__STATIC_FORCEINLINE						//!< \b Inline attribute alias when __NOOPT__ defined
+#ifdef __STATIC_INLINE
+#undef __STATIC_INLINE
+#endif
+#define __STATIC_INLINE		__STATIC_FORCEINLINE						//!< \b Static \b Inline attribute alias when __NOOPT__ defined
 #endif